test: add unit tests and snapshots for oracle-adapter admin and pause functionality
