+++ contracts/oracle-adapter/src/events.rs
@@ -24,7 +24,7 @@ pub struct Paused {
 #[contractevent]
 #[derive(Clone, Debug, PartialEq)]
 pub struct Unpaused {
-    pub paused: bool,
+    pub by: Address,
 }
 
 #[contractevent]
